refactor: Summary와 dailySummary 의존성 제거
